requirements.txt
langgraph
langchain-openai
python-dotenv
langchain
langchain-community
langchain-chroma
pypdf
langgraph-supervisor
flask
flask-cors
langgraph-checkpoint-redis
langchain-ollama
